Rewrite in python
There was a lot of strange behavior in vimscript version of this plugin.
The on_stdout callback appeared to be called a variable number of times
before the on_exit callback, most likely indicating a race. I tried
refactoring to call the rev-list command sequentially for each plugin in
an async cascade, but for unknown reasons, stdout was blank for about
half of the calls in my setup.

In python, we can use a comparatively simple synchronous function
wrapped in vim.async_call, so that's what I've done. The plugin now
appears to finally work consistently and as expected. Since
async/job.vim is no longer needed, I've removed that as well.